Some refactoring, refresh from test
