+++ src/documents/views.py
@@ -8,14 +8,12 @@ import tempfile
 import urllib
 import zipfile
 from datetime import datetime
-from datetime import timedelta
 from pathlib import Path
 from time import mktime
 from unicodedata import normalize
 from urllib.parse import quote
 from urllib.parse import urlparse
 
-import httpx
 import pathvalidate
 from django.apps import apps
 from django.conf import settings
@@ -2158,87 +2156,3 @@ class TrashView(ListModelMixin, PassUserMixin):
                 doc_ids = [doc.id for doc in docs]
             empty_trash(doc_ids=doc_ids)
         return Response({"result": "OK", "doc_ids": doc_ids})
-
-
-# Outlook https://stackoverflow.com/questions/73902642/office-365-imap-authentication-via-oauth2-and-python-msal-library
-class OauthCallbackView(GenericAPIView):
-    # permission_classes = (AllowAny,)
-
-    def get(self, request, format=None):
-        code = request.query_params.get("code")
-        # Gmail passes scope as a query param, Outlook does not
-        scope = request.query_params.get("scope")
-
-        if code is None:
-            logger.error(
-                f"Invalid oauth callback request, code: {code}, scope: {scope}",
-            )
-            return HttpResponseBadRequest("Invalid request, see logs for more detail")
-
-        if scope is not None and "google" in scope:
-            # Google
-            # Gmail setup guide: https://postmansmtp.com/how-to-configure-post-smtp-with-gmailgsuite-using-oauth/
-            imap_server = "imap.gmail.com"
-            defaults = {
-                "name": f"Gmail OAuth {datetime.now()}",
-                "username": "",
-                "imap_security": MailAccount.ImapSecurity.SSL,
-                "imap_port": 993,
-            }
-
-            token_request_uri = "https://accounts.google.com/o/oauth2/token"
-            client_id = settings.GMAIL_OAUTH_CLIENT_ID
-            client_secret = settings.GMAIL_OAUTH_CLIENT_SECRET
-            scope = "https://mail.google.com/"
-        elif scope is None:
-            # Outlook
-            # Outlok setup guide: https://medium.com/@manojkumardhakad/python-read-and-send-outlook-mail-using-oauth2-token-and-graph-api-53de606ecfa1
-            imap_server = "outlook.office365.com"
-            defaults = {
-                "name": f"Outlook OAuth {datetime.now()}",
-                "username": "",
-                "imap_security": MailAccount.ImapSecurity.SSL,
-                "imap_port": 993,
-            }
-
-            token_request_uri = (
-                "https://login.microsoftonline.com/common/oauth2/v2.0/token"
-            )
-            client_id = settings.OUTLOOK_OAUTH_CLIENT_ID
-            client_secret = settings.OUTLOOK_OAUTH_CLIENT_SECRET
-            scope = "offline_access https://outlook.office.com/IMAP.AccessAsUser.All"
-
-        data = {
-            "code": code,
-            "client_id": client_id,
-            "client_secret": client_secret,
-            "scope": scope,
-            "redirect_uri": "http://localhost:8000/api/oauth/callback/",
-            "grant_type": "authorization_code",
-        }
-        headers = {
-            "Content-Type": "application/x-www-form-urlencoded",
-        }
-        response = httpx.post(token_request_uri, data=data, headers=headers)
-        data = response.json()
-
-        if "error" in data:
-            logger.error(f"Error {response.status_code} getting access token: {data}")
-            return HttpResponseRedirect(
-                "http://localhost:4200/mail?oauth_success=0",
-            )
-        elif "access_token" in data:
-            access_token = data["access_token"]
-            refresh_token = data["refresh_token"]
-            expires_in = data["expires_in"]
-            account, _ = MailAccount.objects.update_or_create(
-                password=access_token,
-                is_token=True,
-                imap_server=imap_server,
-                refresh_token=refresh_token,
-                expiration=timezone.now() + timedelta(seconds=expires_in),
-                defaults=defaults,
-            )
-            return HttpResponseRedirect(
-                f"http://localhost:4200/mail?oauth_success=1&account_id={account.pk}",
-            )
